Revert "Updated api routes, removed jwt auth"
