--- public/gameFiles/v1/styles.css
@@ -1,16 +1,0 @@
-body {
-  margin: 0;
-  overflow: hidden;
-}
-
-canvas {
-  display: block;
-  background-color: black;
-  margin: 0 auto;
-}
-
-#gameCanvas {
-  border: 1px solid #000;
-  margin: 0 auto;
-  display: block;
-}
