+++ public/input.js
@@ -31,5 +31,3 @@ document.addEventListener("keyup", (event) => {
 });
 
 
-
-
